backend/requirements.txt
# Core Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==24.1.0
websockets==12.0

# Computer Vision & Hand Tracking
opencv-python==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4
Pillow==10.4.0

# Audio Processing
SpeechRecognition==3.10.0
pyttsx3==2.90
gTTS==2.3.2
pydub==0.25.1
soundfile==0.12.1
librosa==0.10.1
webrtcvad==2.0.10

# AI & Machine Learning
cerebras-cloud-sdk==1.0.0
sign-language-translator==0.8.1

# I3D Model Support
torch>=2.1.0
torchvision>=0.16.0
scipy==1.11.4
scikit-learn==1.3.2

# TGCN Model Support
torch-geometric>=2.4.0
mediapipe==0.10.14  # Already listed above, but critical for pose extraction

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development Tools
black==23.12.0
isort==5.13.0
flake8==6.1.0